add: 连带删除,多删,角色权限显示名称,admin 不能自删等逻辑

+++ pages/system/user/list.vue
@@ -9,12 +9,14 @@
 				<input class="uni-search" type="text" v-model="query" placeholder="请输入搜索内容" />
 				<button class="uni-button" type="default" size="mini" @click="search">搜索</button>
 				<button class="uni-button" type="default" size="mini" @click="navigateTo('./add')">新增</button>
-				<!-- <button class="uni-button" type="default" size="mini" @click="delTable">批量删除</button> -->
+				<button class="uni-button" type="default" size="mini" @click="delTable">批量删除</button>
 			</view>
 		</view>
 		<view class="uni-container">
-			<uni-clientdb ref="udb" :collection="collectionName" :options="options" :where="where" page-data="replace" :orderby="orderby"
-			 :getcount="true" :page-size="options.pageSize" :page-current="options.pageCurrent" v-slot:default="{data,pagination,loading,error}">
+			<uni-clientdb ref="udb" :collection="collectionName" :options="options" :where="where" field="username,role{role_id,role_name},mobile,email,status,
+register_date"
+			 page-data="replace" :orderby="orderby" :getcount="true" :page-size="options.pageSize" :page-current="options.pageCurrent"
+			 v-slot:default="{data,pagination,loading,error}">
 				<uni-table :loading="loading" :emptyText="error.message || '没有更多数据'" border stripe type="selection"
 				 @selection-change="selectionChange">
 					<uni-tr>
@@ -23,21 +25,26 @@
 						<uni-th align="center">角色</uni-th>
 						<uni-th align="center">手机号</uni-th>
 						<uni-th align="center">邮箱</uni-th>
-						<uni-th align="center">是否启用</uni-th>
+						<uni-th align="center">用户状态</uni-th>
+						<uni-th width="170" align="center">创建时间</uni-th>
 						<uni-th width="204" align="center">操作</uni-th>
 					</uni-tr>
 					<uni-tr v-for="(item,index) in data" :key="index">
 						<uni-td align="center">{{item.username}}</uni-td>
 						<!-- <uni-td align="center">{{item.password}}</uni-td> -->
-						<uni-td align="center">{{item.role ? item.role.join('，') : '-'}}</uni-td>
+						<uni-td align="center">{{item.role ? item.role.map(item => item.role_name).join('、') : '-'}}</uni-td>
 						<uni-td align="center">{{item.mobile}}</uni-td>
 						<uni-td align="center">{{item.email}}</uni-td>
 						<uni-td align="center">{{item.role && item.role.includes('admin') ? '启用' : parseUserStatus(item.status)}}</uni-td>
 						<uni-td align="center">
-							<view v-if="item.role ? item.role.includes('admin') : false">-</view>
+							<uni-dateformat :date="item.
+register_date" :threshold="[0, 0]" />
+						</uni-td>
+						<uni-td align="center">
+							<view v-if="item._id === userInfo._id">-</view>
 							<view v-else class="uni-group">
 								<button @click="navigateTo('./edit?id='+item._id)" class="uni-button" size="mini" type="primary">修改</button>
-								<button @click="confirmDelete(item._id)" class="uni-button" size="mini" type="warn">删除</button>
+								<button @click="confirmDelete(item)" class="uni-button" size="mini" type="warn">删除</button>
 							</view>
 						</uni-td>
 					</uni-tr>
@@ -54,13 +61,16 @@
 <script>
 	const db = uniCloud.database()
 	// 表查询配置
-	const dbCollectionName = 'uni-id-users'
+	const dbCollectionName = 'uni-id-users,uni-id-roles'
 	const dbOrderBy = '' // 排序字段
 	const dbSearchFields = [] // 支持模糊搜索的字段列表
 	// 分页配置
 	const pageSize = 10
 	const pageCurrent = 1
 
+	import {
+		mapState
+	} from 'vuex'
 	export default {
 		data() {
 			return {
@@ -73,6 +83,9 @@
 					pageCurrent
 				}
 			}
+		},
+		computed: {
+			...mapState('user', ['userInfo']),
 		},
 		methods: {
 			getWhere() {
@@ -118,13 +131,32 @@
 			},
 			//批量删除
 			delTable() {
-				this.$refs.udb.remove(this.selectedItems())
+				const ids = this.selectedItems()
+				const currentUserId = this.userInfo._id
+				if (ids.includes(currentUserId)) {
+					uni.showToast({
+						icon: 'none',
+						title: '当前账号不能删除自己',
+						duration: 1500
+					})
+					return
+				}
+				this.$refs.udb.remove(ids)
 			},
 			// 多选
 			selectionChange(e) {
 				this.selectedIndexs = e.detail.index
 			},
-			confirmDelete(id) {
+			confirmDelete(item) {
+				const currentUserId = this.userInfo._id
+				if (currentUserId === item._id) {
+					uni.showToast({
+						icon: 'none',
+						title: '不允许账号删除自己',
+						duration: 1500
+					})
+					return
+				}
 				this.$refs.udb.remove(id)
 			},
 			parseUserStatus(status) {
